requirements.txt
altair==4.1.0
vega==3.4.0
pandas==1.2.3
cython==0.29.22
future==0.17.1
networkx==2.5
psycopg2==2.8.6
pyyaml==5.4.1
scipy==1.6.3
numpy==1.19.2
disjoint-set==0.7.2
matplotlib==3.3.4
cvxopt==1.2.6
